feat(ci): automate release preparation flow
